게시글 삭제 api 작성 (cascade 처리)

+++ src/main/java/kr/co/opgg/datasource/file/File.java
@@ -20,7 +20,7 @@ public class File {
     @Id
     @GeneratedValue
     @Column(name = "FILE_IDX")
-    private Long fileIdx;
+    private Integer fileIdx;
 
     @Column(name = "FILE_NAME")
     private String fileName;
